malaria_detection_project/pyproject.toml
[project]
name = "content"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.135.1",
    "numpy>=2.4.3",
    "pillow>=12.1.1",
    "python-multipart>=0.0.22",
    "scikit-learn>=1.8.0",
    "tensorflow>=2.21.0",
    "uvicorn>=0.41.0",
]
